% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{create_filesystem}
\alias{create_filesystem}
\title{Create filesystem}
\usage{
create_filesystem(internal, public, settings)
}
\arguments{
\item{internal}{Filepath. Folder to hold internal reports.}

\item{public}{Filepath. Folder to hold public reports.}

\item{settings}{Filepath. Folder to hold report settings.}
}
\value{
NULL.
}
\description{
\code{create_filesystem} creates the given folders if they don't already exist.
}
\examples{
 internal_folder = file.path(tempdir(), "internal")
 public_folder = file.path(tempdir(), "public")
 settings_folder = file.path(tempdir(), "settings")

 create_filesystem(
   internal = internal_folder,
   public = public_folder,
   settings = settings_folder
 )
}
