% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{create_public_report_ytd}
\alias{create_public_report_ytd}
\title{Create a YTD public report}
\usage{
create_public_report_ytd(ytd_rates, d_list, config)
}
\arguments{
\item{ytd_rates}{Dataframe. YTD case rates per 100k. Must have columns:
disease, Current_YTD_Rate_per_100k, Avg_5yr_YTD_Rate_per_100k.}

\item{d_list}{Dataframe. List of diseases to use for the report. Must have
columns: EpiTrax_name, Public_name.}

\item{config}{List. Settings to use for report.}
}
\value{
List containing the report name and data.
}
\description{
'create_public_report_ytd' creates a public report for YTD rates.
}
\examples{
ytd_rates <- data.frame(
  disease = c("A","B"),
  Current_YTD_Rate_per_100k = c(12, 34),
  Avg_5yr_YTD_Rate_per_100k = c(10, 30)
)
d_list <- data.frame(
  EpiTrax_name = c("A","B"),
  Public_name = c("Alpha","Beta")
)
config <- list(generate_csvs = TRUE, trend_threshold = 0.15)
create_public_report_ytd(ytd_rates, d_list, config)
}
