% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{format_week_num}
\alias{format_week_num}
\title{Format input EpiTrax data}
\usage{
format_week_num(data)
}
\arguments{
\item{data}{Dataframe. Data to format.}
}
\value{
The formatted data.
}
\description{
'format_week_num' formats the input EpiTrax dataset with month numbers
using the field 'patient_mmwr_week' and filters rows older than five years.
}
\examples{
df <- data.frame(patient_mmwr_year=2020L, patient_mmwr_week=1L, patient_disease="A")
format_week_num(df)
}
