% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{get_internal_disease_list}
\alias{get_internal_disease_list}
\title{Get the internal disease list}
\usage{
get_internal_disease_list(filepath, default_diseases)
}
\arguments{
\item{filepath}{Filepath. Internal disease list CSV file.}

\item{default_diseases}{String vector. List of default diseases to use if the
above file doesn't exist.}
}
\value{
A dataframe containing the diseases to include in the internal report
and possibly the disease groupings.
}
\description{
'get_internal_disease_list' reads the internal list from a given CSV file or
uses the default diseases, if the file doesn't exist.
}
\details{
The provided internal disease list file must contain a column of EpiTrax
disease names (EpiTrax_name) to include in internal reports. It can optionally
contain a column for disease_groupings (Group_name) for reports that group
diseases together. See the example file here:
\code{system.file("sample_data/sample_disease_list.csv", package = "epitraxr")}
}
\examples{
# Using default list (when file doesn't exist)
default_list <- c("Measles", "Chickenpox")
disease_list <- get_internal_disease_list("", default_list)

# Using a disease list file
list_file <- system.file("sample_data/sample_disease_list.csv",
                        package = "epitraxr")
disease_list <- get_internal_disease_list(list_file, default_list)
}
