% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{get_public_disease_list}
\alias{get_public_disease_list}
\title{Get the public disease list}
\usage{
get_public_disease_list(filepath, default_diseases)
}
\arguments{
\item{filepath}{Filepath. Public disease list CSV file.}

\item{default_diseases}{String vector. List of default diseases to use if the
above file doesn't exist.}
}
\value{
A dataframe containing the diseases to include in the public report
and the name to use for each disease in the public report.
}
\description{
'get_public_disease_list' reads the public list from a given CSV file or uses
the default diseases if the file doesn't exist.
}
\details{
The provided public disease list file must contain two columns that map the
EpiTrax disease name to a public-facing name for the public report.
}
\examples{
# Using default list (when file doesn't exist)
default_list <- c("Measles", "Chickenpox")
disease_list <- get_public_disease_list("", default_list)

# Using a disease list file
list_file <- system.file("tinytest/test_files/disease_lists/public_list.csv",
                        package = "epitraxr")
disease_list <- get_public_disease_list(list_file, default_list)
}
