% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{get_report_disease_lists}
\alias{get_report_disease_lists}
\title{Get both internal and public disease lists}
\usage{
get_report_disease_lists(internal_list_fp, public_list_fp, default_diseases)
}
\arguments{
\item{internal_list_fp}{Filepath. Path to internal disease list CSV file.}

\item{public_list_fp}{Filepath. Path to public disease list CSV file.}

\item{default_diseases}{String vector. List of default diseases to use if
either file doesn't exist.}
}
\value{
A list with two elements:
\itemize{
\item internal: Dataframe with EpiTrax_name column
\item public: Dataframe with EpiTrax_name and Public_name columns
}
}
\description{
\code{get_report_disease_lists} is a convenience function that combines
\code{get_internal_disease_list} and \code{get_public_disease_list}.
}
\examples{
# Using default lists (when files don't exist)
default_list <- c("Measles", "Chickenpox")
disease_lists <- get_report_disease_lists("", "", default_list)

# Using disease list files
i_file <- system.file("tinytest/test_files/disease_lists/internal_list.csv",
                       package = "epitraxr")
p_file <- system.file("tinytest/test_files/disease_lists/public_list.csv",
                       package = "epitraxr")
disease_lists <- get_report_disease_lists(
  internal_list_fp = i_file,
  public_list_fp = p_file,
  default_diseases = default_list
)
}
