% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{get_report_diseases_public}
\alias{get_report_diseases_public}
\title{Get the public disease list}
\usage{
get_report_diseases_public(filepath, defaults)
}
\arguments{
\item{filepath}{Filepath. Public disease list CSV file.}

\item{defaults}{String vector. List of default diseases to use if the
above file doesn't exist.}
}
\value{
A dataframe containing the diseases to include in the public report
and the name to use for each disease in the public report. It may also contain
the disease groupings.
}
\description{
'get_report_diseases_public' reads the public list from a given CSV file or uses
the default diseases if the file doesn't exist.
}
\details{
The provided public disease list file must contain two columns that map the
EpiTrax disease name (EpiTrax_name) to a public-facing name (Public_name)
for the public report. It can optionally contain a column for disease_groupings
(Group_name) for reports that group diseases together. See the example file here:
\code{system.file("sample_data/sample_disease_list.csv", package = "epitraxr")}
}
\examples{
# Using default list (when file doesn't exist)
default_list <- c("Measles", "Chickenpox")
disease_list <- get_report_diseases_public("", default_list)

# Using a disease list file
list_file <- system.file("sample_data/sample_disease_list.csv",
                        package = "epitraxr")
disease_list <- get_report_diseases_public(list_file, default_list)
}
