% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_trend}
\alias{get_trend}
\title{Get Trend column of report}
\usage{
get_trend(col1, col2)
}
\arguments{
\item{col1}{List. Current data.}

\item{col2}{List. Historical comparison data.}
}
\value{
Column containing the Trend markers.
}
\description{
'get_trend' compares values of two columns and produces a new column
containing the trend result. The trend is represented by arrows:
up, down, and right arrows for increase, decrease, and no change respectively.
Arrows are given using escape codes to generate Unicode characters.
}
\examples{
get_trend(c(5, 10, 10), c(3, 10, 12))
}
