% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_yrs}
\alias{get_yrs}
\title{Get unique years from a data frame}
\usage{
get_yrs(data)
}
\arguments{
\item{data}{Dataframe. Must contain a 'year' column.}
}
\value{
Integer vector of sorted unique years present in the data.
}
\description{
'get_yrs' extracts and returns the sorted unique years from the 'year' column
of a data frame.
}
\examples{
df <- data.frame(year = c(2020, 2021, 2020, 2022))
get_yrs(df)
}
