% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prep_report_data}
\alias{prep_report_data}
\title{Prepare data for report}
\usage{
prep_report_data(data, report_d_list)
}
\arguments{
\item{data}{Dataframe. Current report data.}

\item{report_d_list}{String vector. Diseases to include in the report.}
}
\value{
Report data with rows for all diseases to report.
}
\description{
'prep_report_data' removes rows from the data that shouldn't appear in the
report and adds rows for diseases that should be in the report, but weren't
in the input dataset. Added rows are filled with 0s.
}
\examples{
df <- data.frame(disease=c("A","B","D"), counts=c(5,7,8))
prep_report_data(df, c("A","C"))
}
