% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{read_report_config}
\alias{read_report_config}
\title{Read in the report config YAML file}
\usage{
read_report_config(config_filepath)
}
\arguments{
\item{config_filepath}{Filepath. Path to report config file.}
}
\value{
a named list with an attribute of 'keys' from the file.
}
\description{
'read_report_config' reads in the config YAML file
}
\examples{
\dontrun{
  config_file <- "path/to/config_file"
  report_config <- read_report_config(config_file)
}
}
