% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{reshape_annual_wide}
\alias{reshape_annual_wide}
\title{Reshape data frame with each year as a separate column}
\usage{
reshape_annual_wide(df, yrs)
}
\arguments{
\item{df}{Dataframe. Data to reshape with years as columns. Must have
columns: disease, year, and counts.}

\item{yrs}{Integer vector. Years to use as columns.}
}
\value{
The reshaped data frame.
}
\description{
'reshape_annual_wide' reshapes a given data frame with diseases for rows and
years for columns.
}
\examples{
df <- data.frame(
  disease = c("A", "A", "B"),
  year = c(2020, 2021, 2020),
  counts = c(5, 7, 8)
)
reshape_annual_wide(df, yrs = c(2020, 2021))
}
