% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{reshape_monthly_wide}
\alias{reshape_monthly_wide}
\title{Reshape data with each month as a separate column}
\usage{
reshape_monthly_wide(data)
}
\arguments{
\item{data}{Dataframe. Must contain columns:
\itemize{
\item \code{disease} (character)
\item \code{month} (integer)
\item \code{counts} (integer)
}}
}
\value{
The reshaped data frame.
}
\description{
'reshape_monthly_wide' reshapes a given data frame with diseases for rows and
months for columns.
}
\examples{
df <- data.frame(
 disease = c("A", "B"),
 month = c(1, 2),
 counts = c(5, 6)
)
reshape_monthly_wide(df)
}
