% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the epitraxr Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{shiny::shinyAppDir()}.}
}
\value{
Starts the execution of the app, printing the port
to the console.
}
\description{
\code{run_app} launches the interactive Shiny web application for
EpiTrax data analysis and report generation. The app provides a user-friendly
interface for uploading data, configuring reports, and generating various
types of disease surveillance reports.
}
\examples{
if (interactive() & requireNamespace("shiny")) {
  run_app()
}
}
