% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_na_0}
\alias{set_na_0}
\title{Set NA values to 0}
\usage{
set_na_0(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
Dataframe with NA values replaced by 0.
}
\description{
'set_na_0' sets NA values to 0 in a data frame.
}
\examples{
df <- data.frame(year = c(2020, NA, 2022))
set_na_0(df)
}
