% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_config}
\alias{validate_config}
\title{Validate config}
\usage{
validate_config(config)
}
\arguments{
\item{config}{Named list.}
}
\value{
A named list with 'keys' corresponding to config options.
}
\description{
\code{validate_config} checks the values of the given config list. If any values
are missing or invalid, they are set to default values and a warning is
issued.
}
\examples{
validate_config(config = list())
}
