% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{validate_data}
\alias{validate_data}
\title{Validate input EpiTrax data}
\usage{
validate_data(data)
}
\arguments{
\item{data}{Dataframe. EpiTrax data to validate.}
}
\value{
The validated data with all unneeded columns removed.
}
\description{
'validate_data' checks the data for expected columns and data types, removes
unneeded columns, and returns the resulting data.
}
\examples{
df <- data.frame(patient_mmwr_year=2020L, patient_mmwr_week=1L, patient_disease="A")
validate_data(df)
}
