% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_filesystem}
\alias{validate_filesystem}
\title{Validate filesystem structure}
\usage{
validate_filesystem(fsys)
}
\arguments{
\item{fsys}{List. Contains paths to report folders with elements:
\itemize{
\item \code{internal}: Folder for internal reports
\item \code{public}: Folder for public reports
}}
}
\value{
NULL if valid, otherwise throws an error.
}
\description{
\code{validate_filesystem} checks that the filesystem structure is valid.
}
\examples{
fsys <- list(
  internal = "internal_reports",
  public = "public_reports"
)
validate_filesystem(fsys)
}
