% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{clear_old_reports}
\alias{clear_old_reports}
\title{Clear out old reports before generating new ones.}
\usage{
clear_old_reports(i_folder, p_folder)
}
\arguments{
\item{i_folder}{Filepath. Folder containing internal reports.}

\item{p_folder}{Filepath. Folder containing public reports.}
}
\value{
The list of old reports that were cleared.
}
\description{
\code{clear_old_reports} deletes reports from previous runs and returns a list of
the reports that were deleted.
}
\examples{
\dontrun{
  clear_old_reports("internal_reports", "public_reports")
}
}
