% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{create_public_report_month}
\alias{create_public_report_month}
\title{Create a monthly cross-section public report}
\usage{
create_public_report_month(cases, avgs, d_list, m, y, config)
}
\arguments{
\item{cases}{Dataframe. Disease case counts for each month and year. Must
have columns: disease, year, month, counts.}

\item{avgs}{Dataframe. Disease case count averages for each month. Must
have columns: disease, Jan, Feb, ...}

\item{d_list}{Dataframe. List of diseases to use for the report. Must have
columns: EpiTrax_name, Public_name.}

\item{m}{Integer. The report month (1-12).}

\item{y}{Integer. The report year.}

\item{config}{List. Settings to use for report.}
}
\value{
List containing the report name and data.
}
\description{
'create_public_report_month' creates a public report for the given month.
}
\examples{
cases <- data.frame(
  disease = c("A","B"),
  year = 2024,
  month = 1,
  counts = c(10,20)
)
avgs <- data.frame(disease = c("A","B"), Jan = c(5,15))
d_list <- data.frame(
  EpiTrax_name = c("A","B"),
  Public_name = c("Alpha","Beta")
)
config <- list(
  current_population = 100000,
  avg_5yr_population = 100000,
  rounding_decimals = 1
)

create_public_report_month(cases, avgs, d_list, 1, 2024, config)
}
