% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_trend}
\alias{get_trend}
\title{Get Trend column of report}
\usage{
get_trend(col1, col2)
}
\arguments{
\item{col1}{List. Current data.}

\item{col2}{List. Historical comparison data.}
}
\value{
Character vector containing the trend labels.
}
\description{
'get_trend' compares values of two columns and produces a new column
containing the trend result. The trend is represented by the strings:
"Elevated" (increase), "Less Than Expected" (decrease), and
"Expected" (no change).
}
\examples{
get_trend(c(5, 10, 10), c(3, 10, 12))
}
