% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_yrs}
\alias{get_yrs}
\title{Get unique years from the data}
\usage{
get_yrs(data)
}
\arguments{
\item{data}{Dataframe. Must contain the column:
\itemize{
\item \code{year} (integer)
}}
}
\value{
Integer vector of sorted unique years present in the data.
}
\description{
'get_yrs' extracts and returns the sorted unique years from the 'year' column
of a data frame.
}
\examples{
df <- data.frame(year = c(2020, 2021, 2020, 2022))
get_yrs(df)
}
