% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_epitrax_data}
\alias{read_epitrax_data}
\title{Read in input EpiTrax data}
\usage{
read_epitrax_data(data_file = NULL)
}
\arguments{
\item{data_file}{Optional filepath. Data file should be a CSV. If this parameter
is NULL, the user will be prompted to choose a file interactively.}
}
\value{
The validated and formatted EpiTrax data from the input file.
}
\description{
'read_epitrax_data' reads EpiTrax data from a CSV, validates and formats it,
then returns the data.
}
\examples{
\dontrun{
# Interactive file chooser:
read_epitrax_data()
}
# Using a file path:
read_epitrax_data(system.file("sample_data/sample_epitrax_data.csv",
                              package = "epitraxr"))
}
