% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{reshape_monthly_wide}
\alias{reshape_monthly_wide}
\title{Reshape data frame with each month as a separate column}
\usage{
reshape_monthly_wide(df)
}
\arguments{
\item{df}{Dataframe. Data to reshape with months as columns.}
}
\value{
The reshaped data frame.
}
\description{
'reshape_monthly_wide' reshapes a given data frame with diseases for rows and
months for columns.
}
\examples{
df <- data.frame(disease=c("A","B"), month=c(1,2), counts=c(5,6))
reshape_monthly_wide(df)
}
