% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitrax.R
\name{setup_epitrax}
\alias{setup_epitrax}
\title{Setup EpiTrax object with configuration and disease lists}
\usage{
setup_epitrax(epitrax_file, config_file, disease_list_files)
}
\arguments{
\item{epitrax_file}{Optional path to the EpiTrax data file. Data file should
be a CSV. If this parameter is NULL, the user will be prompted to choose a
file interactively.}

\item{config_file}{Path to the report configuration file.}

\item{disease_list_files}{List containing filepaths to internal and public
report disease lists.}
}
\value{
An EpiTrax object with configuration and report diseases set.
}
\description{
\code{setup_epitrax} initializes an EpiTrax object with configuration and report
disease lists. It is a convenience function that combines \code{get_epitrax},
\code{epitrax_add_config}, and \code{epitrax_add_report_diseases}.
}
\examples{
data_file <- system.file("sample_data/sample_epitrax_data.csv",
                         package = "epitraxr")
config_file <- system.file("tinytest/test_files/configs/good_config.yaml",
                           package = "epitraxr")
disease_lists <- list(
  internal = system.file("tinytest/test_files/disease_lists/internal_list.csv",
                         package = "epitraxr"),
  public = system.file("tinytest/test_files/disease_lists/public_list.csv",
                       package = "epitraxr")
)

epitrax <- setup_epitrax(
  epitrax_file = data_file,
  config_file = config_file,
  disease_list_files = disease_lists
)
}
