% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{validate_filesystem}
\alias{validate_filesystem}
\title{Validate the filesystem structure}
\usage{
validate_filesystem(fsys)
}
\arguments{
\item{fsys}{List. Contains paths to report folders with elements:
\itemize{
\item internal: Folder for internal reports
\item public: Folder for public reports
\item settings: Folder for settings files
}}
}
\value{
NULL if valid, otherwise throws an error.
}
\description{
\code{validate_filesystem} checks that the filesystem structure is valid.
}
\examples{
fsys <- list(
  internal = "internal_reports",
  public = "public_reports",
  settings = "report_settings"
)
validate_filesystem(fsys)
}
